Safari-609.1.11.1/LayoutTests/storage/websql/statement-success-callback-isolated-world-expected.txt

CONSOLE MESSAGE: line 6: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: line 45: Web SQL is deprecated. Please use IndexedDB instead.
ALERT: undefined
ALERT: PASS: document.body.bar visible in a callback created in this world.
This test tests that the statement success callback is called in the right world.
